==> notebooks/reports/OpenAI/1.0_urls_google.ipynb <==
import os
from googleapiclient.discovery import build
from dotenv import load_dotenv
import pickle
import random
import collections
from tqdm import tqdm

random.seed(333)

load_dotenv()

api_key = os.environ['GOOGLE_API']
cse_id = os.environ['GOOGLE_CSE']

# create a search query
def google_search(search_term, api_key, cse_id, **kwargs):
    
    """
    Creates a search link for the custom Google search.
    """
    
    service = build("customsearch", "v1", developerKey=api_key)
    res = service.cse().list(q=search_term, cx=cse_id, **kwargs).execute()
    return res['items']

# search_results = google_search(palm_species[0], api_key=api_key, cse_id=cse_id)

queries = [
    "",
    "description",
    "diagnosis",
    "attributes"
]

## PALMS

folder = "../../../data/OpenAI/Species/"

palm_species = pickle.load(open(F"{folder}palm_species.pkl", 'rb'))
# palm_species_random = random.sample(palm_species, 50)


try:
    folder = "../../../data/OpenAI/Urls/"
    palm_species_google = pickle.load(open(F"{folder}google_urls_palms.pkl", 'rb'))
    palm_species_urls = pickle.load(open(F"{folder}urls_palms.pkl", 'rb'))
    palms_done = list(palm_species_urls.keys())
except:
    palm_species_urls = collections.defaultdict(list)
    palm_species_google = collections.defaultdict(list)
    palms_done = []

palms_done

for palm in tqdm(palm_species):

    if palm in palms_done:
        print(palm, "done")
        continue

    for query in queries:
        search_query = F"{palm} {query}"

        # print(search_query)

        # Search results (10 per search)
        search_results = google_search(search_query, api_key=api_key, cse_id=cse_id)
        # Google Urls
        palm_species_google[palm] = search_results
        # Just the links
        for result in search_results:
            palm_species_urls[palm].append(result['link'])


folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_urls, f)

with open(F"{folder}google_urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_google, f)


folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_urls, f)

with open(F"{folder}google_urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_google, f)


## CARIBBEAN

folder = "../../../data/OpenAI/Species/"

caribbean_species = pickle.load(open(F"{folder}caribbean_species.pkl", 'rb'))

try:
    folder = "../../../data/OpenAI/Urls/"
    caribbean_species_google = pickle.load(open(F"{folder}google_urls_caribbean.pkl", 'rb'))
    caribbean_species_urls = pickle.load(open(F"{folder}urls_caribbean.pkl", 'rb'))
    caribbean_done = list(caribbean_species_urls.keys())
except:
    caribbean_species_urls = collections.defaultdict(list)
    caribbean_species_google = collections.defaultdict(list)
    caribbean_done = []

caribbean_done

for caribbean in tqdm(caribbean_species):

    if caribbean in caribbean_done:
        print(caribbean, "done")
        continue

    for query in queries:
        search_query = F"{caribbean} {query}"

        # print(search_query)

        # Search results (10 per search)
        search_results = google_search(search_query, api_key=api_key, cse_id=cse_id)
        # Google Urls
        caribbean_species_google[caribbean] = search_results
        # Just the links
        for result in search_results:
            caribbean_species_urls[caribbean].append(result['link'])


folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_urls, f)

with open(F"{folder}google_urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_google, f)

folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_urls, f)

with open(F"{folder}google_urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_google, f)

## PlantNet

folder = "../../../data/OpenAI/Species/"

plantnet_species = pickle.load(open(F"{folder}plantnet_species.pkl", 'rb'))

try:
    folder = "../../../data/OpenAI/Urls/"
    plantnet_species_google = pickle.load(open(F"{folder}google_urls_plantnet.pkl", 'rb'))
    plantnet_species_urls = pickle.load(open(F"{folder}urls_plantnet.pkl", 'rb'))
    plantnet_done = list(plantnet_species_urls.keys())
except:
    plantnet_species_urls = collections.defaultdict(list)
    plantnet_species_google = collections.defaultdict(list)
    plantnet_done = []

plantnet_done

for plantnet in tqdm(plantnet_species):

    if plantnet in plantnet_done:
        print(plantnet, "done")
        continue

    for query in queries:
        search_query = F"{plantnet} {query}"

        # print(search_query)

        # Search results (10 per search)
        search_results = google_search(search_query, api_key=api_key, cse_id=cse_id)
        # Google Urls
        plantnet_species_google[plantnet] = search_results
        # Just the links
        for result in search_results:
            plantnet_species_urls[plantnet].append(result['link'])


folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_plantnet.pkl", 'wb') as f:
    pickle.dump(plantnet_species_urls, f)

with open(F"{folder}google_urls_plantnet.pkl", 'wb') as f:
    pickle.dump(plantnet_species_google, f)

folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_plantnet.pkl", 'wb') as f:
    pickle.dump(plantnet_species_urls, f)

with open(F"{folder}google_urls_plantnet.pkl", 'wb') as f:
    pickle.dump(plantnet_species_google, f)
