==> notebooks/reports/OpenAI/3.1_dataframe_to_traits.ipynb <==
import pandas as pd
import json


### Dataframes

root = "../../../data/OpenAI/DataFrames/"

file = "DF_Andrei.csv"
df_Andrei = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')

file = "DF_Daniel.csv"
df_Daniel = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')

file = "DF_Pierre.csv"
df_Pierre = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')


def create_trait_dict(multi_index):
    # Create an empty dictionary
    result_dict = {}

    # Iterate through the MultiIndex object
    for idx in multi_index:
        key = idx[0]
        value = idx[1]
        if key not in result_dict:
            result_dict[key] = []
        # Error from Andre
        if 'NA' in value:
            continue
        result_dict[key].append(value)
    
    return result_dict

root = "../../../data/OpenAI/Traits/"

for (name, df) in zip(
            ["Andrei", "Daniel", "Pierre"], 
            [df_Andrei, df_Daniel, df_Pierre]
            ):
    res = create_trait_dict(df.columns)

    with open(F"{root}{name}.json", 'w') as fp:
        json.dump(res, fp)
