==> notebooks/reports/OpenAI/4.1_description_text_cleaner.ipynb <==
import json
import re
import glob
import collections
from tqdm import tqdm 

def regex_cleaner(string):
    
    cleaners = [
        ("\s+", " "),
    ]

    for (cleaner, replacement) in cleaners:
        string = re.sub(cleaner, replacement, string)

    return string


def json_para_cleaner(json_para_file):

    with open(json_para_file, 'r') as f:
        json_para_file_unclean = json.load(f)

    paragraphs_cleaned = []
    json_para_cleaned = {}

    for species, paragraphs in json_para_file_unclean.items():

        for paragraph in paragraphs:

            # Keep only larger text
            if len(paragraph) > 50:

                paragraph = regex_cleaner(paragraph)
                paragraphs_cleaned.append(paragraph)

                # Drop duplicates
                paragraphs_cleaned = list(set(paragraphs_cleaned))
        
        json_para_cleaned[species] = paragraphs_cleaned

    return json_para_cleaned           
    

folder_description_paragraphs = "../../../data/OpenAI/DescriptionSnippets/Paragraphs/"

json_para_unclean_list = glob.glob(F"{folder_description_paragraphs}*")
json_para_unclean_list.sort()

folder_description_paragraphs_cleaned = "../../../data/OpenAI/DescriptionSnippetsCleaned/Paragraphs/"

for f in tqdm(json_para_unclean_list):

    file_name = F"{f[52:-5]}_cleaned"
    res = json_para_cleaner(f)

    with open(F"{folder_description_paragraphs_cleaned}{file_name}.json", 'w') as fp:
        json.dump(res, fp)
