==> notebooks/reports/OpenAI/5.2_ChatGPT_prompting_per_species_per_trait.ipynb <==
import torch
import re
import os
import json
import pickle
import openai
import pandas as pd
import numpy as np
import glob
from tqdm import tqdm
from dotenv import load_dotenv

### Dotenv

load_dotenv()
openai.api_key = os.environ['OPENAI_API']

### DataFrames

root = "../../../data/OpenAI/DataFrames/"

file = "DF_Andrei.csv"
df_Andrei = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')
df_Andrei_species = list(df_Andrei.index)


file = "DF_Daniel.csv"
df_Daniel = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')\
    .dropna()
df_Daniel_species = list(df_Daniel.index)

file = "DF_Pierre.csv"
df_Pierre = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')
df_Pierre_species = list(df_Pierre.index)

### Trait Dicts

folder_traits = "../../../data/OpenAI/Traits/"

with open(F"{folder_traits}Andrei.json", 'r') as f:
  caribbean_traits_dict = json.load(f)

### Text Data

paragraph_folder = "../../../data/OpenAI/DescriptionSnippets/Paragraphs/"
sentence_folder = "../../../data/OpenAI/DescriptionSnippets/Sentences/"

caribbean_jsons_paras = glob.glob(F"{paragraph_folder}c*")
caribbean_jsons_sents = glob.glob(F"{sentence_folder}c*")

caribbean_jsons_paras.sort()
caribbean_jsons_sents.sort()

### Single Species Testing

with open(caribbean_jsons_paras[22], 'r') as f:
  caribbean_species_para_test = json.load(f)

with open(caribbean_jsons_sents[22], 'r') as f:
  caribbean_species_sent_test = json.load(f)

species = list(caribbean_species_sent_test.keys())[0]
print(species)

text_sent = ' '.join(caribbean_species_sent_test[species])
text_paras = ' '.join(caribbean_species_para_test[species])

text_paras == text_sent

### Prompting per Trait

def create_prompt(paragraph, traits, question):

    prompt = F"""
    Text:
    {paragraph}

    Trait with possible options:
    {traits}

    {question}
    """

    return prompt

def combine_words_with_capital(string):
    # remove non-alphanumeric characters
    string = re.sub(r'[^\w\s]', '', string)  
    # combine words with capitalization
    string = ''.join(word.capitalize() for word in string.split())  
    return string


def query_ChatGPT(prompt):
    messages = [
    {"role": "user", "content": prompt}
    ]

    # Generate a response
    completion = openai.ChatCompletion.create(
        model="gpt-3.5-turbo",
        messages=messages,
    )

    return completion

caribbean_traits_dict = {k: caribbean_traits_dict[k][0:] for k in list(caribbean_traits_dict)[:3]}
caribbean_traits_dict

folder_prompts = "../../../data/OpenAI/PromptsResults/"
folder_species = species.replace(' ', '_')


# For spss in species:

try:
    os.makedirs(F"{folder_prompts}{folder_species}")
except FileExistsError:
    # directory already exists
    pass

question = "Can you select the correct value from the Python trait list, fill in 'NOT SPECIFIED' if you cannot find anything."

for trait, trait_options in tqdm(caribbean_traits_dict.items()):

    # Prompt
    trait_with_options = F"- {trait}: {trait_options}"
    prompt = create_prompt(text_paras, trait_with_options, question)

    result = query_ChatGPT(prompt)
    # # TESTING
    # result = {'Test':
    #           [
    #             "String1", 
    #             "String2",
    #           ]}

    file_name = combine_words_with_capital(trait)
    with open(F"{folder_prompts}{folder_species}/{file_name}.json", 'w') as fp:
        json.dump(result, fp)

result

print(result["choices"][0]["message"]["content"])

print(prompt)
